requirements.txt
pika==1.2.0
redis==3.5.3
gunicorn==20.0.4
python-dotenv==0.15.0
psycopg2-binary==2.8.3
channels==2.4.0
channels-redis==2.4.2
django==2.2.5
djangorestframework==3.10.3
django-filter==2.2.0
django-cleanup==4.0.0
django-cors-headers==3.1.0
django-picklefield==2.0
